backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.1
anthropic==0.40.0
openai==1.57.4
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.10.3
